% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{d2ir}
\alias{d2ir}
\alias{ir2d}
\title{Convert between Geographic Coordinates and ICES Rectangles (copied from geo R package)}
\usage{
d2ir(lat, lon = NULL, useI = FALSE)

ir2d(ir, useI = FALSE)
}
\arguments{
\item{lat}{vector of latitudes, or a list containing \code{lat} and
\code{lon}.}

\item{lon}{vector of longitudes (ignored if \code{lat} is a list).}

\item{useI}{whether to use the letter \sQuote{\samp{I}} in statistical
rectangle codes.}

\item{ir}{ICES rectangle code, e.g. \samp{37F3}}
}
\value{
Vector of strings containing ICES statistical rectangle codes or conversely, center coordinates of rectangles which codes were given as input.
}
\description{
This function converts geographic coordinates to ICES North Atlantic
statistical rectangles and rectangle codes to rectangle center coordinates.
}
\details{
The default \code{useI=FALSE} is in accordance with the prescription in ICES CM77/Gen:3, but \code{useI=TRUE} has been done on occasion.
}
\examples{

d2ir(54.25, 3.5)
d2ir(c(50,60), c(-20,-10))

ir2d(d2ir(54, 3))
## center positions for bottom left and approx top right rects
ir2d("01A0")
ir2d("99M7")
ir2d(c("01A0","99M7"))
## note that ICES CM1977/Gen:3 indicates half-size rects on eastern margin!

}
