% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_catches_link.R
\name{tg_catches_link}
\alias{tg_catches_link}
\title{BioSamplesCatches}
\usage{
tg_catches_link(cn.standardized = FALSE, lowercase = FALSE)
}
\arguments{
\item{cn.standardized}{Boolean, if FALSE (default) retains variable names as
delivered by the webserver otherwise mri-standaridzed variable names are used (not active)}

\item{lowercase}{Boolean, if TRUE, variable names are set to lower case. If FALSE,
names will be consitent with documentation.}
}
\value{
A dataframe
}
\description{
BioSamplesCatches
}
\examples{
df <- tg_catches_link()
}
