% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_expeditions.R
\name{tg_expeditions}
\alias{tg_expeditions}
\title{Expeditions}
\usage{
tg_expeditions(cn.standardized = FALSE, lowercase = FALSE)
}
\arguments{
\item{cn.standardized}{Boolean, if FALSE (default) retains variable names as
delivered by the webserver otherwise mri-standaridzed variable names are used}

\item{lowercase}{Boolean, if TRUE, variable names are set to lower case. If FALSE,
names will be consitent with documentation.}
}
\value{
A dataframe
}
\description{
Expeditions
}
\examples{
df <- tg_expeditions()
}
