% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_expeditions_bio.R
\name{tg_expeditions_bio}
\alias{tg_expeditions_bio}
\title{Expeditions biology}
\usage{
tg_expeditions_bio(species = "mackerel")
}
\arguments{
\item{species}{"mackerel" or "herring"}
}
\value{
A dataframe
}
\description{
Expeditions biology
}
