% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_expeditions_link.R
\name{tg_expeditions_link}
\alias{tg_expeditions_link}
\title{BiosamplesExpeditions}
\usage{
tg_expeditions_link(
  species = "mackerel",
  cn.standardized = FALSE,
  lowercase = FALSE
)
}
\arguments{
\item{species}{"mackerel" or "herring"}

\item{cn.standardized}{Boolean, if FALSE (default) retains variable names as
delivered by the webserver otherwise mri-standaridzed variable names are used}

\item{lowercase}{Boolean, if TRUE, variable names are set to lower case. If FALSE,
names will be consitent with documentation.}
}
\value{
A data frame
}
\description{
BiosamplesExpeditions
}
\examples{
tg_expeditions_link()
}
