% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_out_of_order.R
\name{tg_outoforder}
\alias{tg_outoforder}
\title{List of factories to exclude for certain periods}
\usage{
tg_outoforder(lowercase = FALSE)
}
\arguments{
\item{lowercase}{Boolean, if TRUE, variable names are set to lower case. If FALSE,
names will be consitent with documentation.}
}
\value{
A data frame
}
\description{
List of factories to exclude for certain periods
}
\examples{
tg_outoforder()
}
