% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_recaptured.R
\name{tg_recaptured}
\alias{tg_recaptured}
\title{Catches}
\usage{
tg_recaptured(species = "mackerel", cn.standardized = FALSE)
}
\arguments{
\item{species}{"mackerel" or "herring"}

\item{cn.standardized}{Boolean, if FALSE (default) retains variable names as
delivered by the webserver otherwise mri-standaridzed variable names are used}
}
\value{
A dataframe
}
\description{
Catches
}
