% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_tagged.R
\name{tg_tagged}
\alias{tg_tagged}
\title{Expeditions}
\usage{
tg_tagged(species = "mackerel", cn.standardized = FALSE)
}
\arguments{
\item{species}{"mackerel" or "herring"}

\item{cn.standardized}{Boolean, if FALSE (default) retains variable names as
delivered by the webserver otherwise mri-standaridzed variable names are used}
}
\value{
A dataframe
}
\description{
Expeditions
}
