% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_catches_bio.R
\name{tg_catches_bio}
\alias{tg_catches_bio}
\title{Catches biology}
\usage{
tg_catches_bio(species = "mackerel", cn.standardized = FALSE)
}
\arguments{
\item{species}{"mackerel" or "herring"}

\item{cn.standardized}{Boolean, if FALSE (default) retains variable names as
delivered by the webserver otherwise mri-standaridzed variable names are used}
}
\value{
A dataframe
}
\description{
Catches biology
}
