% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_expeditions_bio.R
\name{tg_expeditions_bio}
\alias{tg_expeditions_bio}
\title{Expeditions biology}
\usage{
tg_expeditions_bio(species = c("mackerel", "herring"))
}
\arguments{
\item{species}{"mackerel", "herring" or both (default)}
}
\value{
A dataframe
}
\description{
Expeditions biology
}
