% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_expeditions_bio.R
\name{tg_expeditions_bio}
\alias{tg_expeditions_bio}
\title{Expeditions biology}
\usage{
tg_expeditions_bio(lowercase = FALSE)
}
\arguments{
\item{lowercase}{Boolean, if TRUE, variable names are set to lower case. If FALSE,
names will be consitent with documentation.}
}
\value{
A dataframe
}
\description{
Expeditions biology
}
\examples{
df <- tg_expeditions_bio()
}
